+++ esp-metadata-generated/src/_generated_esp32c3.rs
@@ -48,6 +48,12 @@ macro_rules! property {
     ("soc.rc_fast_clk_default", str) => {
         stringify!(17500000)
     };
+    ("soc.rc_slow_clock") => {
+        136000
+    };
+    ("soc.rc_slow_clock", str) => {
+        stringify!(136000)
+    };
     ("soc.xtal_frequency") => {
         40
     };